requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.4
opencv-contrib-python==4.9.0.80
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.16.1
pyzbar==0.1.9
requests==2.31.0
urllib3==2.2.1
